move tests inline
